Student Portal/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pypdf
python-docx
